### 2024-09-25 #7280 @kthyng
Welcome back @daniloceano and thanks for your submission! I am looking for some specific items to make sure your submission fits our requirements at a high level (not at the more detailed review level) before moving on to finding an editor or putting this on our waitlist if no relevant editors are available. I'll comment over time as I have a chance to go through them:
