### 2024-09-26 #7280 @daniloceano
Hello @kthyng,

Regarding the list of MISSING DOIs, I am a bit confused as the DOIs were provided and direct to the correct papers. Could you please clarify what needs to be done in this case?

As for the reviewers, I would like to suggest the following:

[Mattia Almansi](https://reviewers.joss.theoj.org/reviewers/557)
[koldunovn](https://reviewers.joss.theoj.org/reviewers/859)
[Harald von Waldow](https://reviewers.joss.theoj.org/reviewers/3879)
[Einara Zahn](https://github.com/einaraz)
[amylu00](https://github.com/amylu00)

Best regards,

Danilo
