### 2024-09-26 #7280 @kthyng
@daniloceano The note about DOIs is requesting they be present in your .bib file so they come through the references list. I don't see any DOIs in there: https://github.com/daniloceano/CycloPhaser/blob/joss-submission/paper/paper.bib.
