### 2024-09-26 #7280 @kthyng
@daniloceano comments:
* I see you have docs that say notes about your software. Do you show what it does somewhere? For example, you process vorticity data: https://cyclophaser.readthedocs.io/en/latest/usage.html#example-processing-vorticity-data-from-era5 but it's hard to understand what exactly your software does without some pictures to support it. Would you consider adding this?
* You have two summaries in your paper
* the in line references have parentheses
* Can the package be used in the northern hemisphere now or just in theory?
