### 2024-09-26 2bb083d
Fixed doubled summary, inline references and DOIs
