+++ paper/paper.md
@@ -28,11 +28,6 @@ date: "2024-09-18"
 bibliography: paper.bib
 ---
 
-
-# Summary
-
-CycloPhaser is an open-source Python package designed to detect and analyze extratropical cyclone life cycles using central relative vorticity data. By automating the classification of key cyclone phases, such as intensification, decay, and maturity, CycloPhaser enables efficient processing of large datasets with minimal computational overhead. The package provides customizable options for pre-processing data through noise-reducing filters and smoothing techniques, ensuring accurate phase detection. Initially developed for use in the Southern Hemisphere, CycloPhaser can also be applied to the Northern Hemisphere with minor adjustments. Its flexible design makes it adaptable to various data types, including sea level pressure and geopotential height, broadening its utility for cyclone studies. CycloPhaser can support researchers in atmospheric sciences by enabling in-depth analysis of cyclone life cycles, improving model validation, and offering insights into the impacts of climate change on cyclone behavior.
-
 # Summary
 
 CycloPhaser is a Python package designed to detect and analyze extratropical cyclone life cycles from central relative vorticity data. It enables researchers in meteorology and atmospheric sciences to automatically identify key stages of cyclone development, such as intensification, decay, and mature phases, using the vorticity series and its derivatives. By leveraging vorticity data, CycloPhaser helps scientists study cyclones across various regions and timeframes, contributing to improved understanding of cyclone energetics and behavior.
@@ -41,19 +36,19 @@ CycloPhaser is a Python package designed to detect and analyze extratropical cyc
 
 Extratropical cyclones are key features of the climate system. In South America, they are especially important due to the presence of cyclogenesis hotspots in southeast Brazil (SE-BR), the La Plata River basin (LA-PLATA), and southeastern Argentina (ARG) [@gramcianinov2019properties]. These cyclones can cause extreme precipitation, intense winds, high sea waves, and landslides, significantly impacting communities [@de2021ocean; @cardoso2022synoptic; @gramcianinov2023impact; @de2024extreme]. Understanding their temporal and spatial development and evolution is crucial for improving forecasts, ultimately aiding in the adoption of mitigation and adaptation strategies.
 
-Accurately identifying the regions where cyclones are positioned throughout their distinct life cycle stages remains a significant challenge in atmospheric sciences. Seminal works by [@bjerknes1922life], [@shapiro1990fronts], and [@neiman1993life] described extratropical cyclone life cycles in terms of structural changes and large-scale dynamics. However, these classifications were based on manual analysis of satellite imagery and synoptic charts, limiting their applicability to large datasets with multiple cyclone cases. Recent research has sought to objectively define cyclone life cycle stages using techniques such as normalizing the life cycle duration [@schemm2018during; @rudeva2007climatology] or bisecting the cycle into "intensification" and "decay" phases by focusing on periods before and after peak vorticity or the lowest central pressure [@dacre2009spatial; @trigo2006climatology; @azad2014vorticity; @booth2018extratropical; @michaelis2017changes]. While these approaches support the study of cyclone intensification and decay, they tend to overlook critical phases such as the incipient stage — where environmental dynamics are still adapting to the developing low-level disturbance and surface isobars are not yet fully closed. Additionally, they treat the mature phase as a single time step, failing to account for the possibility that it may encompass multiple time steps during which the cyclone exhibits homogeneous features.
[…]
