+++ paper/paper.bib
@@ -3,14 +3,16 @@
   title={Life cycle of cyclones and the polar front theory of atmospheric circulation},
   author={Bjerknes, Jakob and Solberg, Halvor},
   year={1922},
-  publisher={Grondahl}
+  publisher={Grondahl},
+  doi={https://doi.org/10.1175/1520-0493(1922)50<468:JBAHSO>2.0.CO;2}
 }
 
 @book{shapiro1990fronts,
   title={Fronts, jet streams and the tropopause},
   author={Shapiro, Melvyn A and Keyser, Daniel},
   year={1990},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-1-944970-33-8_10}
 }
 
 @article{neiman1993life,
@@ -20,7 +22,8 @@
   volume={121},
   number={8},
   pages={2153--2176},
-  year={1993}
+  year={1993},
+  doi={10.1175/1520-0493(1993)121<2153:tlcoae>2.0.co;2}
 }
 
 @article{schemm2018during,
@@ -30,7 +33,8 @@
   volume={99},
   number={1},
   pages={149--165},
-  year={2018}
+  year={2018},
+  doi={10.1175/bams-d-16-0261.1}
 }
 
 @article{rudeva2007climatology,
@@ -40,7 +44,8 @@
   volume={135},
   number={7},
   pages={2568--2587},
-  year={2007}
+  year={2007},
+  doi={10.1175/mwr3420.1}
 }
 
 
@@ -52,7 +57,8 @@
   number={1},
   pages={99--115},
   year={2009},
-  publisher={American Meteorological Society}
+  publisher={American Meteorological Society},
+  doi={10.1175/2008mwr2491.1}
 }
 
 @article{azad2014vorticity,
@@ -63,7 +69,8 @@
   number={9},
   pages={3109--3128},
   year={2014},
-  publisher={American Meteorological Society}
+  publisher={American Meteorological Society},
+  doi={10.1175/jas-d-13-0267.1}
 }
 
 @article{booth2018extratropical,
@@ -74,18 +81,8 @@
   number={16},
   pages={8647--8654},
   year={2018},
-  publisher={Wiley Online Library}
-}
-
-@article{booth2018extratropical,
-  title={Extratropical cyclone precipitation life cycles: A satellite-based analysis},
-  author={Booth, James F and Naud, Catherine M and Jeyaratnam, Jeyavinoth},
-  journal={Geophysical Research Letters},
-  volume={45},
-  number={16},
-  pages={8647--8654},
-  year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2018gl078977}
 }
 
 @article{michaelis2017changes,
@@ -96,7 +93,8 @@
   number={17},
   pages={6905--6925},
   year={2017},
-  publisher={American Meteorological Society}
+  publisher={American Meteorological Society},
+  doi={10.1175/jcli-d-16-0697.1}
 }
 
 @article{couto2024new,
@@ -107,7 +105,8 @@
   number={10},
   pages={3568--3588},
   year={2024},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/joc.8539}
 }
 
 @article{virtanen2020scipy,
@@ -118,7 +117,8 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={https://doi.org/10.1038/s41592-019-0686-2}
 }
 
 
@@ -129,7 +129,8 @@ journal={Climate Dynamics},
 volume={53},
[…]
