+++ .circleci/config.yml
@@ -28,7 +28,7 @@ jobs:
 
       - run:
           name: Install package (build_test)
-          command: pipenv install dist/cyclophaser-1.8.1-py3-none-any.whl
+          command: pipenv install dist/cyclophaser-1.8.2-py3-none-any.whl
 
       - run:
           name: Install pytest (build_test)
