### 2024-09-26 #7280 @daniloceano
@kthyng

Thank you for the feedback.

- I have updated the documentation to provide more clarity on what the software does. There is now an Overview section explaining the main purpose of CycloPhaser and a concise description of how it detects cyclone life cycle phases. Additionally, I have added a figure illustrating the CycloPhaser methodology, which visually supports the process of detecting the life cycle stages of cyclones, offering a clearer understanding of the software's capabilities.
- I have deleted the extra summary (oops!)
- I have fixed the inline references and addressed the missing DOIs in the .bib file.
- Regarding Northern Hemisphere support: Yes, CycloPhaser can be used for the Northern Hemisphere as long as the original vorticity series is multiplied by -1 to account for the hemisphere difference in the cyclonic vorticity sign.
