### 2024-09-27 #7280 @daniloceano
A quick update: a have also included the information regarding the northern hemisphere on the documentation overview.
