### 2024-10-02 #7280 @kthyng
@daniloceano The rest of the docs still need to be updated to say that the software works in the northern hemisphere — they seem to indicate that it doesn't.
