+++ docs/conf.py
@@ -26,3 +26,5 @@ exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
 
 html_theme = 'alabaster'
 html_static_path = ['_static']
+
+extensions = ['sphinx.ext.autodoc']
