+++ docs/usage.rst
@@ -1,92 +1,107 @@
+.. _usage:
+
+Usage Guide
+===========
+
+.. contents::
+   :local:
+   :depth: 2
+   
 Arguments and Parameters for determine_periods
 ----------------------------------------------
 
-- **series** (list): A list of vorticity values.
-- **x** (list, optional): Temporal range or other labels for the series. Default is None.
-- **plot** (str, optional): Path to save plots. If provided, plots will be generated and saved. Default is False.
-- **plot_steps** (str, optional): Path to save step-by-step didactic plots. If provided, detailed stepwise plots will be generated. Default is False.
-- **export_dict** (str, optional): Path to save periods as a CSV dictionary. If provided, the periods will be exported to CSV. Default is False.
-- **output_directory** (str, optional): Directory for saving output files (e.g., plots, CSV). Default is './'.
-- **use_filter** (bool, optional): Whether to apply the Lanczos filter to vorticity. Default is 'auto'.
-- **replace_endpoints_with_lowpass** (int, optional): Window length for replacing endpoints with a lowpass filter. Default is 24.
-- **use_smoothing** (bool, optional): Whether to apply Savitzky-Golay smoothing to the filtered vorticity. Default is 'auto'.
-- **use_smoothing_twice** (bool, optional): Whether to apply Savitzky-Golay smoothing twice. Default is 'auto'.
-- **savgol_polynomial** (int, optional): Polynomial order for Savitzky-Golay smoothing. Default is 3.
-- **cutoff_low** (float, optional): Low-frequency cutoff for Lanczos filter. Default is 168.
-- **cutoff_high** (float, optional): High-frequency cutoff for Lanczos filter. Default is 48.
+- **series**: (list) A list of vorticity values to be analyzed.
+- **x**: (list, optional) Temporal range or labels corresponding to the series. This list must be the same length as the vorticity `series`. Default is None.
+- **plot**: (str or bool, optional) Path for saving generated plots. Set to `False` to disable plotting. Default is False.
+- **plot_steps**: (str or bool, optional) Path for saving step-by-step didactic plots. Set to `False` to disable step-wise plotting. Default is False.
+- **export_dict**: (str or bool, optional) Path for exporting the detected periods as a CSV file. Set to `False` to skip exporting. Default is False.
+- **use_filter**: (str or bool, optional) Whether to apply a Lanczos filter to the vorticity data. Specify a window length as an integer, or set to 'auto' to automatically adapt based on the data. Default is 'auto'.
+- **replace_endpoints_with_lowpass**: (int, optional) Replace the endpoints of the series with a lowpass filter. Specify the window length. Default is 24.
+- **use_smoothing**: (str or bool, optional) Apply Savgol smoothing to the vorticity data. Specify the window length, or use 'auto' to adapt based on data length. Default is 'auto'.
+- **use_smoothing_twice**: (str or bool, optional) Apply Savgol smoothing twice for additional noise reduction. Same options as `use_smoothing`. Default is 'auto'.
+- **savgol_polynomial**: (int, optional) Polynomial order for Savgol smoothing. Default is 3.
+- **cutoff_low**: (float, optional) Low-frequency cutoff for the Lanczos filter. Default is 168.
+- **cutoff_high**: (float, optional) High-frequency cutoff for the Lanczos filter. Default is 48.
+
+
+Example Usage
+-------------
+
+Below is an example of using the CycloPhaser package with default options. The function will generate plots and a CSV file that contains detected cyclone life cycle phases.
+
+.. code-block:: python
+
+   from cyclophaser import determine_periods
+   import pandas as pd
+
+   # Load test data
+   track_file = 'tests/test.csv'
[…]
