+++ .circleci/config.yml
@@ -28,7 +28,7 @@ jobs:
 
       - run:
           name: Install package (build_test)
-          command: pipenv install dist/cyclophaser-1.8.5-py3-none-any.whl
+          command: pipenv install dist/cyclophaser-1.8.6-py3-none-any.whl
 
       - run:
           name: Install pytest (build_test)
