+++ paper/paper.bib
@@ -4,7 +4,7 @@
   author={Bjerknes, Jakob and Solberg, Halvor},
   year={1922},
   publisher={Grondahl},
-  doi={https://doi.org/10.1175/1520-0493(1922)50<468:JBAHSO>2.0.CO;2}
+  doi={10.1175/1520-0493(1922)50<468:JBAHSO>2.0.CO;2}
 }
 
 @book{shapiro1990fronts,
@@ -118,7 +118,7 @@
   pages={261--272},
   year={2020},
   publisher={Nature Publishing Group},
-  doi={https://doi.org/10.1038/s41592-019-0686-2}
+  doi={10.1038/s41592-019-0686-2}
 }
 
 
