### 2024-10-03 #7280 @daniloceano
> @daniloceano The rest of the docs still need to be updated to say that the software works in the northern hemisphere — they seem to indicate that it doesn't.

Thank you for the feedback! I have now updated the documentation to clarify that the software works for both the Southern and Northern Hemispheres.

Additionally, I've also made further improvements to the documentation to ensure clarity and ease of use, and also corrected the wrong DOIs.

I hope the updates address your concerns.
