### 2024-10-06 #7280 @observingClouds
Hi @daniloceano,
While I am searching for reviewers, could you please:
- check the formatting of the references `@bjerknes1922life, @shapiro1990fronts, and @neiman1993life` in l. 26, l. 40, l. 66, Fig 1 caption
- add the missing reference in l. 34
- ensure that the references keep their capital letters in the title, `{{}}` can be used to do this in the references file, e.g. it should be `The Vorticity Budgets of North Atlantic Winter Extratropical Cyclone Life Cycles in MERRA Reanalysis` and not `The vorticity budgets of north atlantic winter extratropical cyclone life cycles in MERRA reanalysis`

Thank you!
