### 2024-10-06 bc6ec72
Fixing references and citations.
