+++ paper/paper.md
@@ -36,17 +36,17 @@ CycloPhaser is a Python package designed to detect and analyze extratropical cyc
 
 Extratropical cyclones are key features of the climate system. In South America, they are especially important due to the presence of cyclogenesis hotspots in southeast Brazil (SE-BR), the La Plata River basin (LA-PLATA), and southeastern Argentina (ARG) [@gramcianinov2019properties]. These cyclones can cause extreme precipitation, intense winds, high sea waves, and landslides, significantly impacting communities [@de2021ocean; @cardoso2022synoptic; @gramcianinov2023impact; @de2024extreme]. Understanding their temporal and spatial development and evolution is crucial for improving forecasts, ultimately aiding in the adoption of mitigation and adaptation strategies.
 
-Accurately identifying the regions where cyclones are positioned throughout their distinct life cycle stages remains a significant challenge in atmospheric sciences. Seminal works by `@bjerknes1922life`, `@shapiro1990fronts`, and `@neiman1993life` described extratropical cyclone life cycles in terms of structural changes and large-scale dynamics. However, these classifications were based on manual analysis of satellite imagery and synoptic charts, limiting their applicability to large datasets with multiple cyclone cases. Recent research has sought to objectively define cyclone life cycle stages using techniques such as normalizing the life cycle duration [@schemm2018during; @rudeva2007climatology] or bisecting the cycle into "intensification" and "decay" phases by focusing on periods before and after peak vorticity or the lowest central pressure [@dacre2009spatial; @trigo2006climatology; @azad2014vorticity; @booth2018extratropical; @michaelis2017changes]. While these approaches support the study of cyclone intensification and decay, they tend to overlook critical phases such as the incipient stage — where environmental dynamics are still adapting to the developing low-level disturbance and surface isobars are not yet fully closed. Additionally, they treat the mature phase as a single time step, failing to account for the possibility that it may encompass multiple time steps during which the cyclone exhibits homogeneous features.
+Accurately identifying the regions where cyclones are positioned throughout their distinct life cycle stages remains a significant challenge in atmospheric sciences. Seminal works by Bjerknes et al. [@bjerknes1922life], Shapiro and Keyser [@shapiro1990fronts], and Neiman et al. [@neiman1993life] described extratropical cyclone life cycles in terms of structural changes and large-scale dynamics. However, these classifications were based on manual analysis of satellite imagery and synoptic charts, limiting their applicability to large datasets with multiple cyclone cases. Recent research has sought to objectively define cyclone life cycle stages using techniques such as normalizing the life cycle duration [@schemm2018during; @rudeva2007climatology] or bisecting the cycle into "intensification" and "decay" phases by focusing on periods before and after peak vorticity or the lowest central pressure [@dacre2009spatial; @trigo2006climatology; @azad2014vorticity; @booth2018extratropical; @michaelis2017changes]. While these approaches support the study of cyclone intensification and decay, they tend to overlook critical phases such as the incipient stage — where environmental dynamics are still adapting to the developing low-level disturbance and surface isobars are not yet fully closed. Additionally, they treat the mature phase as a single time step, failing to account for the possibility that it may encompass multiple time steps during which the cyclone exhibits homogeneous features.
 
[…]
