+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @book{bjerknes1922life,
-  title={Life cycle of cyclones and the polar front theory of atmospheric circulation},
+  title={{Life cycle of cyclones and the polar front theory of atmospheric circulation}},
   author={Bjerknes, Jakob and Solberg, Halvor},
   year={1922},
   publisher={Grondahl},
@@ -8,7 +8,7 @@
 }
 
 @book{shapiro1990fronts,
-  title={Fronts, jet streams and the tropopause},
+  title={{Fronts, jet streams and the tropopause}},
   author={Shapiro, Melvyn A and Keyser, Daniel},
   year={1990},
   publisher={Springer},
@@ -16,7 +16,7 @@
 }
 
 @article{neiman1993life,
-  title={The life cycle of an extratropical marine cyclone. Part I: Frontal-cyclone evolution and thermodynamic air-sea interaction},
+  title={{The life cycle of an extratropical marine cyclone. Part I: Frontal-cyclone evolution and thermodynamic air-sea interaction}},
   author={Neiman, Paul J and Shapiro, MA},
   journal={Monthly Weather Review},
   volume={121},
@@ -27,7 +27,7 @@
 }
 
 @article{schemm2018during,
-  title={When during their life cycle are extratropical cyclones attended by fronts?},
+  title={{When during their life cycle are extratropical cyclones attended by fronts?}},
   author={Schemm, Sebastian and Sprenger, Michael and Wernli, Heini},
   journal={Bulletin of the American Meteorological Society},
   volume={99},
@@ -38,7 +38,7 @@
 }
 
 @article{rudeva2007climatology,
-  title={Climatology of cyclone size characteristics and their changes during the cyclone life cycle},
+  title={{Climatology of cyclone size characteristics and their changes during the cyclone life cycle}},
   author={Rudeva, Irina and Gulev, Sergey K},
   journal={Monthly Weather Review},
   volume={135},
@@ -50,7 +50,7 @@
 
 
 @article{dacre2009spatial,
-  title={The spatial distribution and evolution characteristics of North Atlantic cyclones},
+  title={{The spatial distribution and evolution characteristics of North Atlantic cyclones}},
   author={Dacre, Helen Francis and Gray, Suzanne Louise},
   journal={Monthly Weather Review},
   volume={137},
@@ -62,7 +62,7 @@
 }
 
 @article{azad2014vorticity,
-  title={The vorticity budgets of North Atlantic winter extratropical cyclone life cycles in MERRA reanalysis. Part I: Development phase},
+  title={{The vorticity budgets of North Atlantic winter extratropical cyclone life cycles in MERRA reanalysis. Part I: Development phase}},
   author={Azad, Roohollah and Sorteberg, Asgeir},
   journal={Journal of the Atmospheric Sciences},
   volume={71},
@@ -74,7 +74,7 @@
 }
 
 @article{booth2018extratropical,
-  title={Extratropical cyclone precipitation life cycles: A satellite-based analysis},
+  title={{Extratropical cyclone precipitation life cycles: A satellite-based analysis}},
   author={Booth, James F and Naud, Catherine M and Jeyaratnam, Jeyavinoth},
   journal={Geophysical Research Letters},
   volume={45},
@@ -86,7 +86,7 @@
 }
 
 @article{michaelis2017changes,
-  title={Changes in winter North Atlantic extratropical cyclones in high-resolution regional pseudo--global warming simulations},
+  title={{Changes in winter North Atlantic extratropical cyclones in high-resolution regional pseudo--global warming simulations}},
[…]
