### 2024-10-06 #7280 @daniloceano
Hi @observingClouds,

Regarding the references, I initially followed the journal guidelines using `@citation` for direct citations. Since it wasn’t working as expected, I applied a workaround: "Bjerknes et al. (Bjerknes and Solberg 1922), Shapiro and Keyser (Shapiro and Keyser 1990), and Neiman et al. (Neiman and Shapiro 1993)". Would you suggest any alternative, or does this seem like the best approach?

I’ve also addressed the missing references and capitalization issues—everything should be correct now.

Thank you!
