### 2024-10-07 #7280 @observingClouds
Hi @daniloceano,
Thanks for your quick response.

Regarding the capitalization of your references, I see now at closer inspection of your `paper.bib` that the titles are also not capitalized in there. Since that is the source we read the title from, I ask you to adjust the titles their to the version used of the journal. Compare for example your title entry for https://doi.org/10.1175/JAS-D-13-0267.1 with the one on the journals webpage.

In text citations should be created with `@author:2001` and those in parenthesis with `[@author:2001]`. Please note that the **`** are only here for better formatting and should not be used in the paper.md. I think that was the issue why it didn't work for you beforehand. Please adjust accordingly.

Thanks!
