### 2024-10-07 370d4fd
Fixing citations and references captalization
