### 2024-10-07 #7280 @daniloceano
Hi @observingClouds,

Thank you for the clarification. I’ve now followed your suggestion for the in-text citations and adjusted them using the correct format. I’ve also reviewed all the capitalizations in the references and ensured they match the journal’s version, as per your guidance.

Thanks again for your support!
