### 2024-10-17 #7363 @editorialbot
[REVIEW]: CycloPhaser: A Python Package for Detecting Extratropical Cyclone Life Cycles
**Submitting author:** @daniloceano (<a href="http://orcid.org/0000-0003-4121-7583">Danilo Couto de Souza</a>)
**Repository:** https://github.com/daniloceano/CycloPhaser
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 1.9.4
