### 2024-10-18 daniloceano/CycloPhaser#38 @stella-bourdin
[JOSS Review] Include dependencies in installation
Hello,
This issue arises from the JOSS review ongoing here: https://github.com/openjournals/joss-reviews/issues/7363
The installation of the package goes well, but the dependencies are not automatically installed. Can you include it in the installation?
