### 2024-10-18 daniloceano/CycloPhaser#39 @stella-bourdin
[JOSS review] Make test/example file accessible from the package
Hello,
This issue arises from the JOSS review ongoing here: https://github.com/openjournals/joss-reviews/issues/7363

It would be helpful if the example csv file in `tests/test.csv` could be accessed as `cyclophaser.example_file`or something similar.
In the current state of things, the example script in https://cyclophaser.readthedocs.io/en/latest/usage.html cannot be run without first downloading the file in one's own folder, or changing the path to the installation path of the package.

One way to do so is to add something alike the below in `__init__.py`:
```
import pathlib
here = pathlib.Path(__file__).parent
testdata_dir = here / "example_data"
example_file = str(testdata_dir / "test.csv")
```

Another way is to look into `pkg_resources`.
