### 2024-10-18 daniloceano/CycloPhaser#40 @stella-bourdin
[JOSS Review] Documentation example not functionnal
Hello,
This issue arises from the JOSS review ongoing here: https://github.com/openjournals/joss-reviews/issues/7363

The example provided in the documentation (https://cyclophaser.readthedocs.io/en/latest/usage.html) does not work because:
Already raised:
1. Dependencies were not installed when installing the package (https://github.com/daniloceano/CycloPhaser/issues/38)
2. Test file path is not working (https://github.com/daniloceano/CycloPhaser/issues/39)
Additional issue:
3. Column name in the example file is wrong: column `min_zeta_850`does not exist in the example file. Switching to `min_max_zeta_850`works.
Just changing this name in the example once 1 & 2 are treated would fix it.
