### 2024-10-18 #7363 @stella-bourdin
Hi,
Thank you for your submission. I have gone through the paper and ran the documented test. Here are my comments on some points of the review checklists, besides functionality issues raised as issues on the repo itself (links above).

## General checks
### Authorship
Only the first author appear in the commits. Please make sure that authorship follows [journal guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship). I am aware these can include author that did not directly committed into the code, and that you are responsible for the choice of final author list, so this is just a reminder.

## Documentation
### Statement of need
Procedure overview page is great. I would suggest making a similar one for the statement of need (which I supposed can be a copy of the statement of need in the paper).
### Example provided
Besides functionality issues, here are some more comments regarding the example:
* A brief introduction to what is done in this example would be appreciated. In particular, what is the data that is loaded. Help us understand what we are doing here.
* For the output figure, please precise what the different zeta variables correspond to
* In the legend, the residual is green like decay instead of grey.
### API
* Must the `series` argument in `determine_period` be in a specific unit? In any case, I recommend specifying whether or not in the documentation of the function.

## Paper
### Summary, statement of need and state of the field
Very clear and well written
### Features
I suggest adding a first paragraph about what are the input requirements to give context.
Current paragraph 1 lacks clarity, but it would be helped by adding a first paragraph as suggested above.
Current paragraph 2: I suggest citing the paper from the top of the paragraph. e.g. "Thresholds for phase detection were rigorously calibrated **in Couto de Souza et al. (2024)** using a representative set of cyclone tracks, ensuring accurate phase identification while filtering out noise."
Current paragraph 3: It is not clear whether the user can actually plug in an SLP or wind time series and it would work just as well, or whether it would require additional adaptations.

I will next go through a more thorough testing of the code functionalities, and test other data.
Stella
