+++ setup.py
@@ -3,7 +3,7 @@ from setuptools import setup, find_packages
 with open('README.md', encoding='utf-8') as f:
     long_description = f.read()
 
-VERSION = '1.8.6'
+VERSION = '1.8.7'
 DESCRIPTION = 'Determine phases from extratropical cyclone life cycle'
 # LONG_DESCRIPTION = 'This script processes vorticity data, identifies different phases of the cyclone \
     # and plots the identified periods on periods.png and periods_didatic.png'
@@ -18,7 +18,33 @@ setup(
     author_email="danilo.oceano@gmail.com",
     license='MIT',
     packages=find_packages(),
-    install_requires=[],
+    install_requires=[
+        "cmocean==4.0.3",
+        "contourpy==1.3.0",
+        "cycler==0.12.1",
+        "fonttools==4.53.1",
+        "iniconfig==2.0.0",
+        "kiwisolver==1.4.7",
+        "matplotlib==3.9.2",
+        "numpy==2.1.1",
+        "packaging==24.1",
+        "pandas==2.2.2",
+        "pillow==10.4.0",
+        "pluggy==1.5.0",
+        "pyparsing==3.1.4",
+        "pytest==8.3.3",
+        "python-dateutil==2.9.0.post0",
+        "pytz==2024.2",
+        "scipy==1.14.1",
+        "setuptools==72.1.0",
+        "six==1.16.0",
+        "tzdata==2024.1",
+        "wheel==0.44.0",
+        "xarray==2024.9.0"
+    ],
+    package_data={
+        'cyclophaser': ['example_data/example_file.csv'],
+    },
     keywords=['cyclone', 'vorticity', 'meteorology', 'atmospherical sciences'],
     classifiers= [
         "Development Status :: 3 - Alpha",
