+++ cyclophaser/__init__.py
@@ -1,1 +1,5 @@
 from .determine_periods import determine_periods
+import pathlib
+
+here = pathlib.Path(__file__).parent
+example_file = str(here / "example_data" / "example_file.csv")
