+++ .circleci/config.yml
@@ -2,7 +2,7 @@ version: 2.1
 jobs:
   build_test:
     docker:
-      - image: cimg/python:3.11.0
+      - image: cimg/python:3.12.3
     steps:
       - checkout 
 
