+++ docs/requirements.txt
@@ -1,1 +1,47 @@
-sphinx==7.2.6
+alabaster==0.7.16
+babel==2.16.0
+certifi==2024.8.30
+charset-normalizer==3.4.0
+cmocean==4.0.3
+contourpy==1.3.0
+cycler==0.12.1
+distlib==0.3.9
+docutils==0.20.1
+filelock==3.16.1
+fonttools==4.54.1
+idna==3.10
+imagesize==1.4.1
+iniconfig==2.0.0
+Jinja2==3.1.4
+kiwisolver==1.4.7
+MarkupSafe==3.0.2
+matplotlib==3.9.2
+numpy==2.1.2
+packaging==24.1
+pandas==2.2.3
+pillow==11.0.0
+pipenv==2024.1.0
+platformdirs==4.3.6
+pluggy==1.5.0
+Pygments==2.18.0
+pyparsing==3.2.0
+pytest==8.3.3
+python-dateutil==2.9.0.post0
+pytz==2024.2
+requests==2.32.3
+scipy==1.14.1
+setuptools==75.1.0
+six==1.16.0
+snowballstemmer==2.2.0
+Sphinx==7.2.6
+sphinxcontrib-applehelp==2.0.0
+sphinxcontrib-devhelp==2.0.0
+sphinxcontrib-htmlhelp==2.1.0
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==2.0.0
+sphinxcontrib-serializinghtml==2.0.0
+tzdata==2024.2
+urllib3==2.2.3
+virtualenv==20.27.0
+wheel==0.44.0
+xarray==2024.9.0
