### 2024-10-18 1598c23
fixing CircleCi fconfig file
