### 2024-10-18 daniloceano/CycloPhaser#40 @daniloceano
Hi @stella-bourdin,

Thank you again for your feedback. I’ve addressed the final issue regarding the column name in the example. The correct column name is min_max_zeta_850, and I’ve updated the usage documentation accordingly to reflect this change.

Since the other two issues (dependencies installation and test file path) were already resolved, everything should now be in order.

Feel free to test the updated example, and let me know if anything else comes up!

Best regards,
Danilo
