### 2024-10-18 daniloceano/CycloPhaser#39 @daniloceano
Hi @stella-bourdin,

Thank you for your suggestion! We’ve updated the package to include the example CSV file, now named example_file.csv, which can be accessed directly via the package as cyclophaser.example_file. This way, users won’t need to manually download or adjust the path. We’ve also updated setup.py to include the example_file.csv in the package data and adjusted __init__.py accordingly.

Feel free to try it out, and let us know if you have any further feedback!

Best regards,
Danilo
