### 2024-10-18 daniloceano/CycloPhaser#38 @daniloceano
Hi @stella-bourdin,

Thank you for your feedback! I’ve updated the setup.py to ensure that all dependencies are included in the install_requires section. This way, when you install the package via pip install cyclophaser, all the necessary dependencies should now be installed automatically.

Please let me know if you encounter any further issues, and feel free to test it out!

Best regards,
Danilo
