### 2024-10-18 #7363 @daniloceano
@stella-bourdin, thank you for your thorough review. Below, I will address your comments and the steps I have taken to address them:

---

### **Authorship**

- **Comment**: Only the first author appears in the commits. Please make sure that authorship follows journal guidelines. I am aware that these can include authors that did not directly commit to the code, so this is just a reminder.

- **Response**: I appreciate the reminder. I was the only one working directly on the code, but the other researchers involved contributed significantly to the conceptualization of the algorithm, provided datasets, and collaborated with me in checking the outputs. Their contributions were essential for ensuring that the package produces realistic and useful results.

---

### **Documentation**

#### **Statement of Need**
- **Comment**: The procedure overview page is great. I would suggest making a similar one for the statement of need (which can be a copy from the paper).

- **Response**: I have now added a **Statement of Need** section in the documentation, following the suggestion.

#### **Example Provided**
- **Comment**: A brief introduction to what is done in the example would be appreciated. In particular, explain what data is loaded and clarify what we are doing here.

- **Response**: A brief introduction explaining the dataset used and the purpose of the example has been added to the **Usage** section of the documentation.

- **Comment**: For the output figure, please clarify what the different `zeta` variables correspond to.

- **Response**: I simplified the output by removing the extra `zeta` variables in the figure for better clarity.

- **Comment**: In the legend, the residual phase is green like decay instead of grey.

- **Response**: The color for the residual phase has been updated to grey as originally intended.

---

I am currently working on the remaining issues you raised in the repository and will continue to address them one by one. Once all the adjustments are complete, I will notify you and provide the updated code and paper.

Thank you again for your invaluable feedback. I look forward to continuing to improve the paper and the package based on your suggestions.

Best regards,
Danilo
