### 2024-10-19 #7363 @daniloceano
@stella-bourdin following with your appointments:

### **API:**

> _"Must the series argument in `determine_period` be in a specific unit? In any case, I recommend specifying whether or not in the documentation of the function."_

I have clarified in the documentation that the `series` argument does not need to be in a specific unit. While the function was designed for vorticity data, it should theoretically work with other meteorological fields like sea level pressure (SLP) or wind. However, this hasn't been fully tested yet. This update has been made both in the **API documentation** and the **Usage Guide** to make it clear to users.
