+++ paper/paper.md
@@ -44,12 +44,15 @@ The pioneering work by @couto2024new was the first to offer a comprehensive anal
 
 # Features
 
+CycloPhaser requires a time series of relative vorticity data as its primary input. The input data should represent the minimum central relative vorticity of extratropical cyclones, ideally at a specified pressure level (e.g., 850 hPa). This vorticity series, which is typically derived from atmospheric reanalysis datasets, serves as a proxy for cyclone intensity. The series must be provided as a one-dimensional list or array of vorticity values, which will be analyzed to detect key life cycle stages of the cyclone. Additional optional input includes a time or label series for the corresponding vorticity values. Users should ensure the time series corresponds to regular time steps (e.g., hourly or 6-hourly data).
+
 The program includes optional pre-processing steps, such as applying a Lanczos filter to remove noise from the series and a Savitzky-Golay filter for smoothing, ensuring sinusoidal patterns in the data for phase detection. Key cyclone phases — intensification, decay, and mature — are identified through peaks and valleys in the vorticity time series. The intensification phase spans from a vorticity peak to the next valley, while the decay phase covers the opposite. The mature phase is defined as the period between a vorticity valley and neighboring derivative peaks. The pre-processing steps, as well as peaks and valleys detection in the vorticity series, are computed using Scipy's package [@virtanen2020scipy].
 
-Thresholds for phase detection were rigorously calibrated using a representative set of cyclone tracks, ensuring accurate phase identification while filtering out noise. CycloPhaser also includes a residual phase to account for tracking anomalies, such as post-decay re-intensification without returning to maturity. A post-processing step further refines the phase boundaries by correcting gaps and isolating single time-step phases. Finally, the incipient stage is detected by missing labels in the series or by selecting the initial time steps. More details are discussed in @couto2024new.
+Thresholds for phase detection were rigorously calibrated in @couto2024new using a representative set of cyclone tracks, ensuring accurate phase identification while filtering out noise. CycloPhaser also includes a residual phase to account for tracking anomalies, such as post-decay re-intensification without returning to maturity. A post-processing step further refines the phase boundaries by correcting gaps and isolating single time-step phases. Finally, the incipient stage is detected by missing labels in the series or by selecting the initial time steps. More details are discussed in @couto2024new.
+
 
 ![Representative example of a cyclone life cycle exhibiting an incipient-intensification-mature-decay configuration.](life-cycle.png)
 
-Although the package was initially devised for detecting life cycle phases using relative vorticity, it can be applied to any time series used as a proxy for cyclone detection, such as sea level pressure and geopotential height. Also, the program was designed for use in the Southern Hemisphere, but it can be applied to Northern Hemisphere vorticity series by multiplying by minus one.
+Although the package was initially devised for detecting life cycle phases using relative vorticity, it could potentially be applied to other time series used as proxies for cyclone detection, such as sea level pressure (SLP) or wind data. However, it has not yet been explicitly tested for these variables. Additionally, the program was designed for use in the Southern Hemisphere, but it can be applied to Northern Hemisphere vorticity series by multiplying the data by minus one, as the input data currently needs to be negative.
 
 # References
