### 2024-10-19 #7363 @daniloceano
@stella-bourdin

#### Features:
- I have added a first paragraph to the **features** section describing the input requirements to provide more context for the user.
- For **paragraph 2**, I followed your suggestion and cited Couto de Souza et al. (2024) at the beginning of the paragraph to clarify that thresholds for phase detection were calibrated based on a representative set of cyclone tracks.
- For **paragraph 3**, I clarified that, although the package was initially developed for vorticity, it could potentially be applied to other variables, such as SLP or wind data. However, I have noted that it has not been explicitly tested for these variables, and the input data currently needs to be negative.

Thank you again for your careful review and valuable suggestions. I look forward to your further feedback as I address these points.

Best regards,
Danilo
