### 2024-10-21 daniloceano/CycloPhaser#38 @stella-bourdin
Hi, I can confirm the installation now include the dependencies, and I was able to run `pip install cyclophaser` and then import the package successfully.
