### 2024-10-23 #7363 @stella-bourdin
Hi,

Thank you for going through this quickly! I am now able to install the program and run the example smoothly.
Some more comments as I continue to test the package (and look for every way to break it - sorry).

### Paper
I am not sure where I can find the updated of the paper (the link above seem to correspond to the previous version). Should it not appear on this thread? (@observingClouds ?)

### Documentation
Although minor, I feel like adding one example plot for a canonical case on the GitHub and readthedocs first page would greatly help the users directly visualizing what the package does.

### Tests

It seems to me that the test only checks that the program runs without error. It could be helpful to also check that the output is what was expected, in case a change does not break the program, but affects the output.

### `determine_periods` functionality and documentation
These are thoughts on improving the package that include more or less essential changes. Feel free to discard those you do not wish to implement.
* `series`: It would make the package more easy to use if it could accept more flexible inputs. I'm thinking numpy arrays in particular, but also pandas series or xarray datarrays as the user is I think most likely to have loaded its data with pandas or xarray. However, the error message if the type of input is not right is quite clear, so it is up to you.
* `x` : It is not exactly clear what type this can be and what is the expected behaviour based on this. It seems like I can provide it with numpy or pandas datetime with similar behaviour. I am not sure what it is doing when I provide a list of ints.
* Pre-processing options: It is not clear what is the units of these arguments? Is it in number of time steps? In hours? It would help to provide some recommendations on how to tune these parameters best to our data (should we tune it to the frequency, for example?)
   * `use_smoothing`:
     * an error is raised if it is below `savgol_polynomial` (I think), it would be good to precise that it must be greater.
     * Doc says it can be boolean, but False and True raise errors (because it reads them as 0 or 1 which are below `savgol_polynomial = 3`. Is it actually possible to deactivate?
   * `use_filter` : I seem to obtain the same result whether I activate it or not, is it normal? Why does the yellow line not fit the grey one after I removed the filter? Is there still some kind of normalization?
* Hemisphere support: I would argue in favor of making this an argument. It is as simple as changing the sign of the input, but as user it makes it more friendly. That would help for using NH data, but also if one want to use wind speed data where apex is necessarily the maxima in both hemisphere.

Maybe a test could be added in case the pre-processing parameters choice leads to the apparition of "spurious oscillation" at the beginning and end of the series, making the user aware he might need to adapt the parameters. I suppose this could be done by checking the delta between the first and second point compared to the order of magnitude of the input values.

### Testing with other data
Here are some observations as I apply the package to a set of my tropical cyclones tracks I have lying around.
I find it slightly confusing why sometimes the first part is incipient, and other times when it is not obvious to me they display different behaviours.

In this case, there is a residual in the middle of the time series. Is it normal?

In this case, there is a gap with nothing identified in the middle. Is it normal?

Sometimes the identified mature stage seems offset compared to the actual series. In the example below, the red part is well below the obvious vorticity minimum. How can I fix this?

Using wind data instead of vorticity, I seem able to get satisfactory results. When using sea-level pressure data, it gets weirder. I think it is because there is some normalization going on at some point, making it more difficult to read the plots.
