### 2024-10-23 #7363 @observingClouds
@stella-bourdin I just generated the latest version of the manuscript (see above). We do not do this automatically, but one can generate the latest version by kindly asking our bot 😄
