+++ tests/test_determine_periods.py
@@ -1,40 +1,46 @@
 from cyclophaser.determine_periods import determine_periods
 import pandas as pd
+import numpy as np
+import xarray as xr
+import warnings
+
+def load_expected_csv(filepath):
+    return pd.read_csv(filepath)
 
 def test_determine_periods_with_options():
+    # Suppress specific warning in this test
+    warnings.filterwarnings("ignore", message="Detected potential spurious oscillations at the series boundaries.*")
     # Read the data from the CSV file
     track_file = 'tests/test.csv'
     track = pd.read_csv(track_file, parse_dates=[0], delimiter=';', index_col=[0])
-    series = track['min_max_zeta_850'].tolist()
-    x = track.index.tolist()
+    series_list = track['min_max_zeta_850'].tolist()
+    x_list = track.index.tolist()
+    
+    # Convert data to other types
+    series_np = np.array(series_list)
+    series_pd = pd.Series(series_list, index=x_list)
+    series_xr = xr.DataArray(series_list, coords=[x_list], dims="time")
 
-    # Options for ERA5 processing
-    options_era5 = {
-        "plot": 'test_ERA5',
-        "plot_steps": 'test_steps_ERA5',
-        "export_dict": 'test_ERA5',
-        "use_filter": 'auto',
-        "replace_endpoints_with_lowpass": 24,
-        "use_smoothing": 'auto',
-        "use_smoothing_twice": 'auto',
-        "savgol_polynomial": 3,
-        "cutoff_low": 168,
-        "cutoff_high": 48,
-        "threshold_intensification_length": 0.075,
-        "threshold_intensification_gap": 0.075,
-        "threshold_mature_distance": 0.125,
-        "threshold_mature_length": 0.03,
-        "threshold_decay_length": 0.075,
-        "threshold_decay_gap": 0.075,
-        "threshold_incipient_length": 0.4
-    }
+    # Define file paths for the expected output files
+    expected_csv_default = 'tests/expected_default.csv'
+    expected_csv_no_filter = 'tests/expected_no_filter.csv'
+    output_csv_default = 'tests/test_dict_default'
+    output_csv_no_filter = 'tests/test_dict_default_no_filter'
 
-    # Call the determine_periods function with options for ERA5
-    result_era5 = determine_periods(series, x=x, **options_era5)
-    assert isinstance(result_era5, pd.DataFrame), "Result should be a DataFrame for ERA5 options."
+    # Test with default parameters and check against expected CSV
+    determine_periods(series_pd, export_dict=output_csv_default)
+    output_df_default = load_expected_csv(f'{output_csv_default}.csv') 
+    expected_df_default = load_expected_csv(expected_csv_default)
+    assert output_df_default.equals(expected_df_default), "Default test failed: Output does not match expected"
 
-    # Options for basic processing without filtering
-    options_basic = {
+    # Test with default parameters without filtering and check against expected CSV
+    determine_periods(series_pd, use_filter=False, export_dict=output_csv_no_filter)
+    output_df_no_filter = load_expected_csv(f'{output_csv_no_filter}.csv')
+    expected_df_no_filter = load_expected_csv(expected_csv_no_filter)
+    assert output_df_no_filter.equals(expected_df_no_filter), "No-filter test failed: Output does not match expected"
+
+    # Test input types (basic functionality check)
+    options = {
         "plot": False,
         "plot_steps": False,
         "export_dict": None,
@@ -48,27 +54,25 @@ def test_determine_periods_with_options():
         "threshold_decay_gap": 0.075,
         "threshold_incipient_length": 0.4
     }
-    
-    # Test basic processing
-    result_basic = determine_periods(series, x=x, **options_basic)
[…]
