### 2024-10-28 #7363 @daniloceano
Hi @stella-bourdin

Thank you so much for your thorough review and feedback! I'm glad to hear that the installation and example run smoothly now, and I appreciate the detailed testing and suggestions. This level of detail will certainly help make the package better, and I’m very grateful for your efforts.

---

### Documentation
We agree that a visual example could greatly enhance usability, and we've now added an example plot to both the GitHub README and the ReadTheDocs front page.

---

### Tests
Your suggestion to verify output accuracy is very helpful. We’ve updated the tests to compare the function’s output against expected results in specific cases, ensuring both functionality and data integrity.

---

### `determine_periods` Functionality and Documentation

We appreciate your detailed feedback on `determine_periods`. Here’s how we’ve addressed each of these points:

- **Flexible Input Types for `series`:** `series` now accepts a wider range of input types, including `list`, `numpy.ndarray`, `pandas.Series`, and `xarray.DataArray`. The function converts these types to a compatible format internally, making the tool more versatile without sacrificing clarity.

- **Clarifying `x` Input:** The `x` parameter has been documented to specify that it accepts a `list` or `pandas.DatetimeIndex` with timestamps. We have also clarified the behavior when using integer values in `x`. This input is crucial for setting the periods based on time intervals, so we’ve elaborated on this in the documentation.

- **Pre-Processing Parameter Units and Recommendations:** Each pre-processing option now specifies time step units. We’ve also added guidance for tuning these parameters based on data frequency, with recommendations to help users adjust values like `cutoff_low` and `cutoff_high` depending on whether their data is daily, hourly, or otherwise.

- **`use_smoothing` Error Handling and Documentation:** You were correct; the smoothing window length must be greater than or equal to `savgol_polynomial`. We’ve added validation to check this, plus an option to deactivate smoothing by setting `use_smoothing=False`. This detail is now clear in both the documentation and error messages.

- **`use_filter` Clarifications:** To activate `use_filter` meaningfully, you should pass an integer value for the filter window length rather than using `True`. Setting `use_filter=True` (I am assuming you did that) without a specified window length defaults the window length to 1, which effectively disables filtering. The yellow line in the plots is normalized to provide a clearer visualization of the filtered and unfiltered series in relation to each other. This normalization, applied both to the filtered and the smoothed series, is specifically for visualization purposes and helps to highlight phase transitions more distinctly.

- **Hemisphere Support:** To support data from both hemispheres and various data types (e.g., wind speed, SLP), we’ve added a `hemisphere` argument. Setting this to "northern" automatically inverts the input, while "southern" (the default) maintains the original convention.

- **Detecting Spurious Oscillations:** We implemented a check to detect potential spurious oscillations at the series boundaries, which might indicate that adjustments are needed for certain pre-processing parameters. If detected, the function issues a warning with specific recommendations for tuning the inputs.

---

### Testing with other data

**Incipient Stage Confusion**:
The package relies on duration thresholds for identifying each phase. Differences in detection can result from variations in input data meeting criteria thresholds, particularly for detecting an initial decay or incipient stage. Using the `plot_steps` argument can help visualize each detection stage step-by-step, providing insights into why certain phases are or aren’t detected in specific instances.

**Unexpected Residual Stage or Gaps in the Series**:
For both of these cases, could you provide details about the options used? These patterns aren’t typical or expected behavior. However, they could relate to challenges in input data quality, especially if there are tracking inconsistencies. For example, in the first case, the tracking algorithm may have terminated the cyclone track prematurely, impacting the ability to fully detect the mature stage. Similarly, the second case might reflect a vorticity series with irregularities, possibly due to the influence of spatial filtering or weak systems.

If you’re using TRACK data, it’s worth noting that TRACK sometimes has limitations in capturing consistent cyclone characteristics, leading to possible anomalies. Without supplementary data—such as spatial fields like SLP—detailed analysis is limited, but data quality issues can occasionally propagate into CycloPhaser’s outputs, amplifying detection inconsistencies.

**Offset in the Identified Mature Stage**:
The offset issue in the mature stage may relate to the smoothing and polynomial settings used. By reducing the level of smoothing or increasing the `savgol_polynomial` value, you can achieve a better alignment of detected vorticity minima with the original series. The filtering process smooths the raw data (yellow line), and adjusting these parameters can mitigate shifts and better capture phase transitions.

**Using Wind or Sea-Level Pressure Data**:
If you’re achieving satisfactory results with wind data, that’s a good sign. Maybe you could now apply the `hemisphere` argument?

Can you share a time series plot of the raw SLP data for reference? The only normalization applied occurs in the plotting stage to facilitate visualization, which might influence readability but does not alter the analysis itself.

---

Thank you again for the constructive feedback, which has truly improved the package. Let us know if you encounter any further issues or have additional suggestions!

Best regards,
Danilo
