### 2024-10-29 #7363 @stella-bourdin
Thank you very much for going through these! Here are additional remarks and replies

### Package dependencies
1. There was something else I forgot to mention, which is that the package installation requires very specific package version for the dependencies. This is likely to create conflict with other packages installed in one's environment. I advise requiring specific versions only when absolutely necessary for cyclophaser to work. You can also request package version >= or <= to a given version is there is a functionality that appeared or disappeared at some point. This flexibility will make the package more easy to use. Otherwise, it needs to be installed in its standalone environment where it cannot be used in conjunction with other packages.

### Replies re. functionality and doc
Thank you for carefully going through these and implementing what I suggested.
2. Regarding `x` : I am checking the function docstring, the API documentation and the usage guide but cannot find information about the behaviour when integers are provided.
3. Regarding pre-processing recommendations: Can you point me to where you provided these recommendations?
4. `use_smoothing` : The docstring is better, but does not say that one may deactivate it altogether by setting False.
5. Regarding the spurious oscillations detector: That is great ! However, it seems triggered by situations where I do not understand why, e.g. for the following test:
![image](https://github.com/user-attachments/assets/ac0cd190-6168-4795-96db-a7dc9abe6a3a)

### Testing with other data
6. **Unexpected Residual Stage or Gaps in the Series**: I used the default options (`determine_periods(list(-1 * tracks[i].relative_vorticity.values), x = list(tracks[i].time.values), plot="test_default",)`), but I agree that the one with residual in the middle is a particularly abnormal time series. The one with a gap however, I would say that by eye I can identify two mature periods. But I suppose this is a matter of tuning. Maybe it could be useful to just add a warning in such cases (A gap in the identified periods, suspicious succession of period...), so that the user is aware of potential data quality problems? As one may not be checking every individual time series when applying to a full dataset, it would warn them that caution is required.
7. Using wind data: I did try to apply the hemisphere argument and it works well indeed.
8. Using SLP data: The second SLP time series was indeed very noisy, hence the result. Input data quality was the problem (garbage in, garbage out...)

### Paper
Seems very good to me now. You might want to update the very last sentence about the hemispheres now that you implemented the option.

### Code
I had a look at the code itself. It is clean and well documented throughout.
You might want to transform some of your testing in the end of each module file into proper tests (some of which might already be). Automated tests do not need to be only for the main functionality but can also be for intermediary steps. But it is up to you.
There is also a chunk of commented imports in `determine_period.py` that you might want to clean out. But I'm being really picky here...
