### 2024-10-29 3877d95
Updated last paragraph for relfeting new hemisphere feature.
