### 2024-10-29 #7363 @daniloceano
Thank you for the feedback!

I’ve updated the requirements.txt file to increase flexibility in dependency versions, specifying only the minimum versions necessary for cyclophaser to function effectively. Testing compatibility with every specific version of each dependency, however, would be impractical due to the extensive range of versions involved. To ensure stability, core dependencies are tested across recent major versions in our CI setup.

---

Regarding `x`: Thank you for pointing that out! I've updated the documentation to clarify the behavior when `x` is provided as a list of integers. In these cases, the function detects phases between the series' time steps rather than actual dates or times, which can be especially useful for data without precise timestamps or relative time steps. Apologies for missing this detail in my previous responses, and thanks again for helping ensure the documentation is as clear as possible!

---

Regarding the pre-processing recommendations, these are detailed in the documentation for each parameter. Here it is:
- **use_filter**: (str or int, optional) Apply a Lanczos filter to the `series`. Specify a window length as an integer to customize or use `'auto'` for adaptive length based on dataset size (half of series length). **Units**: Time steps. **Default**: 'auto'.
  **Recommendation**: If using relative vorticity series, turn off `use_filter` if the tracking procedure already applies spatial filtering to avoid over-filtering and signal loss. For hourly ERA5 data, `'auto'` is typically effective, though this may need adjustment for different temporal resolutions and spatial resolutions. Use smoothing also if noise levels are too high.

- **replace_endpoints_with_lowpass**: (int, optional) Applies a lowpass filter to smooth the series endpoints, which helps stabilize edge effects during filtering. Specify the window length. **Units**: Time steps. **Default**: 24.
  **Recommendation**: For hourly relative vorticity data, a 24-hour (24 time steps) setting is effective. Adjust this based on the temporal and spatial resolution of the original data, especially if using data with higher spatial resolution.

- **use_smoothing**: (str, int, optional) Apply Savgol smoothing to filtered vorticity data. Set to `'auto'` to automatically choose an appropriate window length, or provide an integer window length directly. **Note**: The specified window length must be an odd number and greater than or equal to `savgol_polynomial`. Set `use_smoothing=False` to deactivate. **Units**: Time steps. **Default**: 'auto'.
  **Recommendation**: This setting is sensitive to the length of the time series. The `'auto'` setting uses a window length approximately 1/4 of the series length for series >8 days; otherwise, it uses about 1/2. For lower-noise data, this value can be decreased, and for higher-noise data, increase it accordingly.

- **use_smoothing_twice**: (str, int, optional) Apply a second pass of Savgol smoothing for further noise reduction. This uses similar parameters to `use_smoothing`. **Default**: 'auto'.
  **Recommendation**: This should be a gentler smoothing than the initial `use_smoothing`. The `'auto'` setting applies half the window length used in the first pass.

- **savgol_polynomial**: (int, optional) Polynomial order for Savgol smoothing, which must be less than or equal to the window length specified in `use_smoothing` and `use_smoothing_twice`. **Default**: 3.
  **Recommendation**: Higher values retain sharper peaks and more detailed features but can increase noise; lower values provide a smoother output that may oversmooth finer details. For noisier data, a lower polynomial value is preferable, and for cleaner data, a higher value helps preserve more details.

- **cutoff_low**: (float, optional) Low-frequency cutoff for the Lanczos filter, designed for data with hourly resolution. **Units**: Time steps. **Default**: 168.
  **Recommendation**: Set this to the equivalent of 7 days in time steps to filter out planetary wave influences on vorticity.

- **cutoff_high**: (float, optional) High-frequency cutoff for the Lanczos filter, suitable for reducing high-frequency noise in hourly data. **Units**: Time steps. **Default**: 48.
  **Recommendation**: Set this to the equivalent of 2 days in time steps to effectively filter out mesoscale influences on vorticity.

---

Regarding the spurious oscillations detector: This might be happening because the vorticity delta between the first time steps is much higher than for the rest of the series. This can trigger the detector, but in these cases, the warning can generally be ignored as it’s a conservative check.

---

In response to the issues with residual stages or gaps in the series, we’ve added warnings that detect and alert users if residual periods appear in the middle of a time series or if there are unclassified time steps (gaps). These warnings indicate potential data quality concerns and suggest adjusting pre-processing options, which may help resolve such issues. This update should aid in identifying and addressing unusual behaviors automatically when processing larger datasets without needing to review each individual time series manually.

---

Thanks for confirming that the hemisphere adjustment works well for wind data, and I appreciate the additional feedback on the SLP data. As you noted, input quality does indeed play a big role,

Regarding the code review, thank you for the kind words! I agree with your observation about testing; however, I find that focusing on testing the main functionality, especially for determine_periods, provides clearer debugging and maintenance benefits in this case (at least for me). I've also cleaned up the commented imports in determine_periods.py as suggested.

---

We’ve updated the documentation in the paper to reflect the new `hemisphere` option and clarified its intended use cases, particularly for detecting cyclone phases in both vorticity and alternative series like SLP and wind speed.

---

Thank you once again for the detailed review—it’s made a big difference in refining the package.
