### 2024-11-06 #7363 @stella-bourdin
Regarding the pre-processing recommendations, I do not see them neither in the API doc [here](https://cyclophaser.readthedocs.io/en/latest/api.html), nor in the code docstring [here](https://github.com/daniloceano/CycloPhaser/blob/master/cyclophaser/determine_periods.py). Am I missing something?
