### 2024-11-06 #7363 @daniloceano
Hi @stella-bourdin,

Thanks for catching that typo! I’ve fixed the "CycloPhaser" typo in the documentation.

Regarding the pre-processing recommendations, they’re actually located in the [Usage Guide](https://cyclophaser.readthedocs.io/en/latest/usage.html#customizing-filtering) rather than the API doc. I wanted to keep the API documentation as straightforward as possible, while the Usage Guide offers a more thorough explanation of how to customize filtering.

Let me know if you have any more questions!
