### 2024-11-07 #7363 @freemansw1
Apologies for my delay in getting started, I've just come off of some travel. Initial thoughts are below before I try to break the package using a variety of other data.

## Functionality

### Installation
I wasn't able to install `requirements.txt` using conda-forge, but I was able to install through `pip`. Not worth changing, just wanted to note. Some of the libraries aren't available through conda-forge/conda/mamba.

### Functionality
The example works as described, and the test passes.

## Documentation
### Statement of Need
After the previous reviewer, the authors include a statement of need in the documentation. I think this largely meets the mark, but the homepage of the documentation is not as helpful from this end as I think it can be. I think a reference to this paper or the one in JOC would be useful here. It also needed to be clarified to me how or where one would get a vorticity timeseries, although this becomes clearer the further into the documentation that you go.

I think, ultimately, the best way to resolve my concern here would be to elaborate much more on the homepage about what this package provides an average user. Make it clear that this package is the first that allows translation from tracked ET cyclone vorticity (or potentially other) information to lifecycle stage.

### Installation Instructions
I would have liked instructions on how to install the latest and/or in development package. This is an easy fix. Note also my comment about conda/mamba above. The scientific software world seems split between the two at the moment.

### Example usage
The only example uses pre-packaged data from the package itself. I think showing what this data looks like (or how to generate it by oneself) would be useful. The note on data frequency is useful, but not clear enough. I would have assumed, before reading into the documentation deeply, that if I provided a pd.Series object indexed by a DatetimeIndex, the package would automatically calculate the appropriate values, etc.

I think making it clear that one needs to have the vorticity already from a tracking package (either the one already provided by the author or from a different package) in the examples or on the homepage is a necessary change. That was not immediately clear to me.

### Functionality documentation
I would have liked to have seen the other publicly accessible functions documented with docstrings.

## Software paper
I think the paper is in overall good shape.
