+++ cyclophaser/find_stages.py
@@ -3,6 +3,24 @@ import pandas as pd
 
 def find_mature_stage(df, **args_periods):
 
+    """
+    Identifies and marks the mature stage in the cyclone life cycle based on the 
+    given thresholds for mature distance and length.
+
+    Args:
+        df (pd.DataFrame): DataFrame containing vorticity data with columns for 
+            'z_peaks_valleys' and 'periods'.
+        **args_periods: Variable length argument list containing period-specific 
+            thresholds, including:
+            - 'threshold_mature_distance' (float): Factor to calculate mature 
+              start and end distances from z valleys.
+            - 'threshold_mature_length' (float): Minimum length for a mature 
+              stage as a fraction of the total series length.
+
+    Returns:
+        pd.DataFrame: Updated DataFrame with 'mature' stages marked in the 
+        'periods' column where applicable.
+    """
     threshold_mature_distance = args_periods['threshold_mature_distance']
     threshold_mature_length = args_periods['threshold_mature_length']
 
@@ -62,6 +80,25 @@ def find_mature_stage(df, **args_periods):
 
 def find_intensification_period(df, **args_periods):
 
+    """
+    Identifies and marks the intensification period in the cyclone life cycle 
+    based on the given thresholds for intensification length and gap.
+
+    Args:
+        df (pd.DataFrame): DataFrame containing vorticity data with columns 
+            for 'z_peaks_valleys' and 'periods'.
+        **args_periods: Variable length argument list containing period-specific 
+            thresholds, including:
+            - 'threshold_intensification_length' (float): Minimum length for an 
+              intensification stage as a fraction of the total series length.
+            - 'threshold_intensification_gap' (float): Maximum gap allowed between 
+              consecutive intensification periods as a fraction of the total 
+              series length.
+
+    Returns:
+        pd.DataFrame: Updated DataFrame with 'intensification' stages marked in the 
+        'periods' column where applicable.
+    """
     threshold_intensification_length = args_periods['threshold_intensification_length']
     threshold_intensification_gap = args_periods['threshold_decay_length']
 
@@ -100,6 +137,24 @@ def find_intensification_period(df, **args_periods):
 
 def find_decay_period(df, **args_periods):
 
+    """
+    Identifies and marks the decay stage in the cyclone life cycle based on the 
+    given thresholds for decay length and gap.
+
+    Args:
+        df (pd.DataFrame): DataFrame containing vorticity data with columns for 
+            'z_peaks_valleys' and 'periods'.
+        **args_periods: Variable length argument list containing period-specific 
+            thresholds, including:
+            - 'threshold_decay_length' (float): Minimum decay length as a fraction 
+              of the total series length.
+            - 'threshold_decay_gap' (float): Maximum gap in decay periods as a 
+              fraction of the total series length.
+
+    Returns:
+        pd.DataFrame: Updated DataFrame with 'decay' stages marked in the 
+        'periods' column where applicable.
+    """    
     threshold_decay_length = args_periods['threshold_decay_length']
     threshold_decay_gap = args_periods['threshold_decay_gap']
 
@@ -140,6 +195,23 @@ def find_decay_period(df, **args_periods):
     return df
 
 def find_residual_period(df):
+    """
+    Identifies and fills the 'residual' period in the cyclone life cycle stages where applicable.
+
+    This function analyzes the 'periods' column in the provided DataFrame and marks
+    the NaN values with 'residual' in specific conditions. If there is only one unique
+    phase present, it fills NaNs after the last block of this phase with 'residual'.
+    For multiple phases, it checks the sequence of phases and determines where 'residual'
[…]
