+++ docs/statement_of_need.rst
@@ -1,35 +1,0 @@
-Statement of Need
-=================
-
-Extratropical cyclones are key features of the climate system. In South America, they are especially important due to the presence of cyclogenesis hotspots in southeast Brazil (SE-BR), the La Plata River basin (LA-PLATA), and southeastern Argentina (ARG) [1]_ [2]_. These cyclones can cause extreme precipitation, intense winds, high sea waves, and landslides, significantly impacting communities [3]_ [4]_ [5]_ [6]_. Understanding their temporal and spatial development and evolution is crucial for improving forecasts, ultimately aiding in the adoption of mitigation and adaptation strategies.
-
-Accurately identifying the regions where cyclones are positioned throughout their distinct life cycle stages remains a significant challenge in atmospheric sciences. Seminal works by Bjerknes and Solberg [7]_, Shapiro and Keyser [8]_, and Neiman and Shapiro [9]_ described extratropical cyclone life cycles in terms of structural changes and large-scale dynamics. However, these classifications were based on manual analysis of satellite imagery and synoptic charts, limiting their applicability to large datasets with multiple cyclone cases. 
-
-Recent research has sought to objectively define cyclone life cycle stages using techniques such as normalizing the life cycle duration [10]_ [11]_ or bisecting the cycle into "intensification" and "decay" phases by focusing on periods before and after peak vorticity or the lowest central pressure [12]_ [13]_ [14]_ [15]_ [16]_. While these approaches support the study of cyclone intensification and decay, they tend to overlook critical phases such as the incipient stage — where environmental dynamics are still adapting to the developing low-level disturbance and surface isobars are not yet fully closed. Additionally, they treat the mature phase as a single time step, failing to account for the possibility that it may encompass multiple time steps during which the cyclone exhibits homogeneous features.
-
-The pioneering work by Couto de Souza et al. [17]_ was the first to offer a comprehensive analysis of extratropical cyclone life cycles, dissecting systems into distinct life cycle phases and enabling the detection of multiple configurations across different systems. This study presents the Python package that facilitated such work. The method allows for an automated classification of cyclone life cycle stages, enabling the efficient processing of large datasets with minimal computational cost. This tool opens new avenues for research, such as analyzing cyclone life cycle behavior in climate change projections, enabling comparisons with present-day climates, and providing insights into how cyclone life cycles may evolve in response to climate variability. Additionally, it offers potential for assisting model validation by comparing the spatial positioning of life cycle phases across different models and reanalysis datasets. The package is both flexible and fully customizable, making it adaptable to a wide range of datasets and research needs.
-
-.. figure:: _images/density_map_Aggregate.png
-   :alt: Yearly cyclone track densities for South American cyclogenesis regions.
-
[…]
