### 2024-11-08 #7363 @daniloceano
Thank you very much for your comprehensive review and insights. We've made several updates based on your feedback:

### Installation and Installation Instructions
We've expanded the installation instructions, adding detailed steps for setting up the latest package. The instructions now address setup in various environments, including options for those using `pip` or a virtual `conda` environment.

### Statement of Need
To address the need for clarity on the package's purpose and the source of input data, we have revised both the README and the documentation homepage. These updates emphasize that CycloPhaser does not generate vorticity series and outline how users can obtain these series, either through available tracking packages or from published databases. As it was no longer needed, we removed the `Statement of Need` documentation section.

We've also added citations for both the paper currently under review and the published study in the JOC to provide users with contextual references.

### Example Usage
To further clarify the input data requirements, we’ve added an example track file format in the documentation. This example demonstrates the structure expected for time, latitude, and longitude columns in the input data.

Regarding your suggestion on data frequency, you’re correct that the package automatically handles `pd.Series` objects indexed by a `DatetimeIndex` (as stated in the usage guide and API documentation). The `x` argument is only required when using a list or array, as the `DatetimeIndex` serves as the time reference for `pd.Series` inputs.

### Functionality Documentation
Additional docstrings have been added for other accessible functions to improve usability and help developers integrate or modify parts of CycloPhaser more easily.

Thank you again!
