### 2025-01-15 #7363 @observingClouds
Happy new 2025! @freemansw1 could you please give a quick update on your progress?
