### 2025-03-29 #7363 @observingClouds
@daniloceano I just read through the manuscript and it looks good to go 🎉

Please create a release of your software and publish it e.g. on zenodo. Once you have a version number and a DOI, please post it here and we can continue with the publication process. See also the [author instructions](https://github.com/openjournals/joss-reviews/issues/7363#issuecomment-2744166874).
