### 2025-04-02 #7363 @kthyng
Hi @daniloceano! I am checking out the version of the package and am a little confused. I don't see the typical "releases" or "tags" on the right hand side of your github repo page, though I do see the one tag at the top. I guess this isn't a problem but I haven't seen it before.

Also I see you've listed your co-authors as Contributors in Zenodo, but do you know you can also edit the author metadata to list out your co-authors in a way that is more similar to what is in the paper?
