### 2025-04-02 #7363 @daniloceano
Hi @kthyng!

Thanks for checking out the package. I’m not sure what happened either, but when I tried to create the tag, I came across something mentioning that this feature is deprecated. I didn’t fully understand it, but it seemed like it might be causing the issue. Regarding Zenodo, I’ve actually already edited the author metadata there to match the list in the paper.
