### 2025-04-03 #7363 @observingClouds
@kthyng, @daniloceano you can get the releases listed on the right by clicking on the "settings-gear" you see as an admin on your repository, e.g.:

Then just check "Releases" and they should be shown.
